pyproject.toml
[project]
name = "curso-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.1",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "wikipedia>=1.4.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
]
